feat: basic rbac support + multi-table inheritance
edx-rbac support has been added, and rules defined to grant the
following permissions for now:

* `subsidy.can_read_subsidies`
* `subsidy.can_read_transactions`
* `subsidy.can_create_transactions`

As we add support for more CRUD endpoints, more permissions may need to
be created.  Barebones viewsets are provided, covering ADR
0002-initial-api-specification, and mapped to the permissions.

This commit also discards the SubscriptionSubsidy model for now,
committing to a single Subsidy model for tracking learner credit.  An
ADR has been added to this effect.

4 migrations are included:

0008 - [edx-rbac] supportng models for edx-rbac.
0009 - [subsidy models] Drop tables (deleting data) to prepare for migration to a single Subsidy model.
0010 - [subsidy models] Recreate single Subsidy model.
0011 - [edx-rbac] data migration to seed feature roles.

Note that 0009 is not functionally reversible, dropped tables are not
recoverable.

ENT-6620