feat: list() and retrieve() transactions with more filtering
Before this change, anybody with learner or higher permissions on any
context could perform any REST operation on transactions not their own.
After this change, requesters must explicitly have admin or operator
privileges on a given enterprise customer in order to list(),
retrieve(), (etc.), transactions that do not have a lms_user_id matching
that of the requester.

ENT-6886