fix: get_or_create_learner_credit_subsidy() not idempotent for test records.
Modify the mentioned function such that, if ``default_internal_only`` is False and an existing subsidy is
found with the given ``reference_id``, all `default_*` arguments are ignored
and this function returns that existing record.
However, when ``default_internal_only`` is True, this function will
simply create a new record, regardless of any existing records
with the same ``reference_id`` (we assume that the reference_id is
essentially meaningless for test subsidy records).
ENT-7799