fix: correct variant is included in summary data for content
**Background**: There's an internal helper method to generate "summary"
data for a piece of content, which is used all over this codebase to
extract relevant bits of metadata from said content. The summary data is
also used verbatim in the enterprise-subsidy content metadata API
endpoint (`/api/v1/content-metadata/<content-identifier>/`). The problem
was that this helper method relied on `additional_metadata` from the
content metadata, but that dict is now deprecated.  Furthermore, when a
course RUN key was passed to this helper method, the result would
incorrectly include the variant_id of the advertised run, potentially a
different run than request.

**This PR**: Change logic to actually fetch the `variant_id` of the
requested course run, and fallback to that of the advertised course run.
Furthermore, it cuts `additional_metadata` out of the loop entirely.

ENT-8520