feat: rollback_transaction() also cancels external fulfillments
This is needed for cases where an external fulfillment has been created,
but the subsequent platform fulfillment fails (e.g. due to transient
networking issues).  Before this PR, the external fulfillment would not
get rolled back, resulting in an "orphaned" fulfillment that allows
learners to effectively redeem content for free.  After this PR, that
loophole is fixed.

ENT-8866