feat: remove check for auto external cancelation
The write_reversals_from_enterprise_unenrollments management command (currently run on a cron) has this conditional (source).
We used the not self.automatic_external_cancellation check to gate this code path on a Django setting. On stage/prod, this setting is set to True, and thus bypasses this conditional check.
Now that stage/prod is always True, we can remove this gating altogether.
ENT-9269