fix: coerce CourseLocator to str also when checking refundability given a course
In all of our current code that I'm aware of, we only ever pass a course
run to unenrollment_can_be_refunded(), but I left a bug in the case
where we pass a course.  I'm not completely sure if we need to support
this branch anymore, but as long as its there we shouldn't leave it
broken.