feat: forced enrollment and client-side GEAG idempotency for "duplicate order" error
In order to mitigate the "duplicate order" error from the GEAG create
enterprise allocation endpoint, we should try a little harder to search
for an Order UUID / external reference ID to attach to the transaction.

First, do not instantly give up just because we recieved a "duplicate
order" error. Then search for the Order UUID in possibly two places:

1. On the error response payload. The ID is unlikely to be here today,
   but it might in the future.
2. On the transaction metadata. The ID is also unlikely to be here in
   general, but can be populated when utilizing ForcedPolicyRedemption.

ENT-10879